Only replace python3 at the beginning of lines in scripts
There's a string replace in the generated bash scripts for various
script jobs. It's there to ensure that the same python interpreter which
was used to generate the script is used in the script's python
invocations.

The current blind replacement is a little too aggressive and will
replace any occurrence of 'python3 ' anywhere in the script, including
URLs and package names. This change makes that replacement happen only
at the beginning of a line.